📁📊 Add UpsetPlot  for each paraphrasing models
This commit introduces dependency code to extract individual paraphrases generated by each of the five selected models within our study from the TPME dataset.
This code ensures the generation of compatible data for each model, tailored specifically for integration with the Intervene platform. The ultimate goal is to facilitate the creation of Upset plot graphs.